requirements.txt
Django
djangorestframework
requests
psycopg2
djangorestframework-simplejwt